requirements.txt
sentence-transformers
PyMuPDF
langchain
numpy
google-generativeai
dataclasses; python_version < '3.7'